Make CommandFilter lambda-able
