+++ jtelegrambotapi-core/src/main/java/com/jtelegram/api/commands/filters/RegexFilter.java
@@ -10,7 +10,7 @@ import java.util.regex.Pattern;
  *
  * @author Nick Robson
  */
-public class RegexFilter extends CommandFilter {
+public class RegexFilter extends AbstractCommandFilter {
 
     private final Pattern pattern;
